+++ 02-use-cases/market-trends-agent/Dockerfile
@@ -1,22 +1,16 @@
-FROM public.ecr.aws/docker/library/python:3.10-slim
+FROM public.ecr.aws/docker/library/python:3.12-slim
 WORKDIR /app
 
-# Install system dependencies if needed
 
 
-# Copy entire project (respecting .dockerignore)
-COPY . .
-
-# Install dependencies
-
-
-
-
+COPY requirements.txt requirements.txt
 # Install from requirements file
-RUN python -m pip install --no-cache-dir -r requirements.txt
+RUN pip install -r requirements.txt
 
 
 
+
+RUN pip install aws-opentelemetry-distro>=0.10.1
 
 
 # Set AWS region environment variable
@@ -28,15 +22,15 @@ ENV AWS_DEFAULT_REGION=us-east-1
 # Signal that this is running in Docker for host binding logic
 ENV DOCKER_CONTAINER=1
 
-
-RUN python -m pip install aws_opentelemetry_distro_genai_beta>=0.1.2
-
-
 # Create non-root user
 RUN useradd -m -u 1000 bedrock_agentcore
 USER bedrock_agentcore
 
 EXPOSE 8080
+EXPOSE 8000
+
+# Copy entire project (respecting .dockerignore)
+COPY . .
 
 # Use the full module path
