+++ 02-use-cases/market-trends-agent/tools/__init__.py
@@ -4,17 +4,21 @@ Market Trends Agent Tools
 This package contains all the tools used by the market trends agent:
 - browser_tool: AgentCore Browser Tool integration for web scraping
 - broker_card_tools: Broker card parsing and market summary generation
+- memory_tools: AgentCore Memory integration for broker profiles and conversation history
 """
 
-from .browser_tool import get_stock_data, search_bloomberg_news, search_news
+from .browser_tool import get_stock_data, search_news
 from .broker_card_tools import parse_broker_profile_from_message, generate_market_summary_for_broker, get_broker_card_template, collect_broker_preferences_interactively
+from .memory_tools import create_memory, extract_actor_id, create_memory_tools
 
 __all__ = [
     'get_stock_data',
-    'search_bloomberg_news', 
     'search_news',
     'parse_broker_profile_from_message',
     'generate_market_summary_for_broker',
     'get_broker_card_template',
-    'collect_broker_preferences_interactively'
+    'collect_broker_preferences_interactively',
+    'create_memory',
+    'extract_actor_id',
+    'create_memory_tools'
 ]
